+++ src/rewrite_eit.c
@@ -326,7 +326,7 @@ void eit_rewrite_new_global_packet(unsigned char *ts_packet, rewrite_parameters_
  * This function copy the rewritten EIT to the buffer. And checks if the EIT was changed so the rewritten version have to be updated
  */
 void eit_rewrite_new_channel_packet(unsigned char *ts_packet, rewrite_parameters_t *rewrite_vars, mumudvb_channel_t *channel,
-		multi_p_t *multi_p, unicast_parameters_t *unicast_vars, void *scam_vars_v,fds_t *fds)
+		unicast_parameters_t *unicast_vars, void *scam_vars_v,fds_t *fds)
 {
 	int i=0;
 	//If payload unit start indicator , we will send all the present EIT for this service, otherwise nothing
@@ -437,7 +437,7 @@ void eit_rewrite_new_channel_packet(unsigned char *ts_packet, rewrite_parameters
 			data_left_to_send=0;
 		}
 		//NOW we fill the channel buffer for sending
-		buffer_func(channel, send_buf, unicast_vars, multi_p, scam_vars_v, fds);
+		buffer_func(channel, send_buf, unicast_vars, scam_vars_v, fds);
 	}
 
 	//We update which section we want to send
