+++ src/unicast_http.h
@@ -225,7 +225,7 @@ int read_unicast_configuration(unicast_parameters_t *unicast_vars, mumudvb_chann
 
 void unicast_data_send(mumudvb_channel_t *actual_channel,  fds_t *fds, unicast_parameters_t *unicast_vars);
 
-
+void init_unicast_v(unicast_parameters_t *unicast_vars);
 
 
 #endif
